requirements.txt
tensorflow==2.18.0
streamlit
numpy
pillow

